pyproject.toml
[project]
name = "millm"
version = "1.0.0"
description = "Mechanistic Interpretability LLM Server with SAE support"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "miLLM Team" }]
keywords = ["llm", "interpretability", "sae", "mechanistic", "transformers"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",

    # Configuration
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # WebSocket
    "python-socketio>=5.12.0",

    # ML/AI
    "torch>=2.5.0",
    "transformers>=4.47.0",
    "huggingface-hub>=0.27.0",
    "bitsandbytes>=0.45.0",
    "accelerate>=1.2.0",
    "safetensors>=0.5.0",

    # SSE for streaming
    "sse-starlette>=2.2.0",

    # Logging
    "structlog>=25.1.0",

    # Utilities
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.21.0",

    # Linting and formatting
    "ruff>=0.9.0",
    "mypy>=1.14.0",

    # Type stubs
    "types-python-dateutil",
]

redis = [
    "redis>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/miLLM/miLLM"
Documentation = "https://github.com/miLLM/miLLM#readme"
Repository = "https://github.com/miLLM/miLLM"
Issues = "https://github.com/miLLM/miLLM/issues"

[project.scripts]
millm = "millm.main:run"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["millm*"]

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "-v",
    "--strict-markers",
    "-ra",
]
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (requires database)",
    "e2e: End-to-end tests (requires full stack)",
    "slow: Slow tests (skipped by default)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# =============================================================================
# Coverage Configuration
# =============================================================================
[tool.coverage.run]
source = ["millm"]
branch = true
omit = [
    "millm/db/migrations/*",
    "tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true
skip_covered = true

# =============================================================================
# Ruff Configuration (Linting & Formatting)
# =============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (needed for FastAPI Depends)
    "B904",   # raise without from (sometimes intentional)
]

[tool.ruff.lint.isort]
known-first-party = ["millm"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG001"]  # Unused function arguments OK in tests

# =============================================================================
# MyPy Configuration
# =============================================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true

[[tool.mypy.overrides]]
module = [
    "transformers.*",
    "huggingface_hub.*",
    "torch.*",
    "bitsandbytes.*",
    "accelerate.*",
    "safetensors.*",
    "socketio.*",
    "structlog.*",
]
ignore_missing_imports = true
